move `JOB_PASSED_SINCE_TICKET_CREATED_LABEL` from `report` into `constants`
